requirements.txt
numpy
torchvision
matplotlib
torch
opencv-python
dlib
scikit-image
pillow